Renamed files to better reflect their searching functions
