+++ app/src/main/java/com/example/lawn_care/ui/search/SearchViewModel.java
@@ -1,14 +1,14 @@
-package com.example.lawn_care.ui.home;
+package com.example.lawn_care.ui.search;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class SearchViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public SearchViewModel() {
         mText = new MutableLiveData<>();
         mText.setValue("This is home fragment");
     }
