+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/navigation_jobList"
-        android:name="com.example.lawn_care.ui.home.HomeFragment"
+        android:name="com.example.lawn_care.ui.search.SearchFragment"
         android:label="@string/title_jobList"
         tools:layout="@layout/fragment_search_jobs" />
